Update generators structure
