--- lib/generators/decorator/templates/decorator_test.rb
@@ -1,1 +1,0 @@
-# frozen_string_literal: true
