+++ lib/generators/dekorator/decorator/templates/decorator.rb
