+++ lib/generators/dekorator/install/install_generator.rb
