+++ lib/generators/dekorator/rspec/decorator_generator.rb
