+++ lib/generators/dekorator/rspec/templates/decorator_spec.rb
