+++ lib/generators/dekorator/test_unit/templates/decorator_test.rb
